requirements.txt
viam-sdk
numpy
pyinstaller
google-api-python-client
torch
torchvision